dev_requirements.txt
django-debug-toolbar==1.10.1
ipython==7.0.1
-r requirements.txt
